requirements.txt
aiohttp~=3.8.1
loguru~=0.6.0
aiogram~=2.20
python-dotenv~=0.20.0
emoji~=1.7.0